add YOLO, MNIST scripts

+++ centralized-deep-learning-darknet/src/parser.c
@@ -1021,7 +1021,8 @@ void save_weights_upto(network *net, char *filename, int cutoff)
     fwrite(&major, sizeof(int), 1, fp);
     fwrite(&minor, sizeof(int), 1, fp);
     fwrite(&revision, sizeof(int), 1, fp);
-    fwrite(net->seen, sizeof(size_t), 1, fp);
+    //fwrite(net->seen, sizeof(size_t), 1, fp);
+    fwrite(net->seen, sizeof(uint64_t), 1, fp);
 
     int i;
     for(i = 0; i < net->n && i < cutoff; ++i){
@@ -1234,7 +1235,8 @@ void load_weights_upto(network *net, char *filename, int start, int cutoff)
     fread(&minor, sizeof(int), 1, fp);
     fread(&revision, sizeof(int), 1, fp);
     if ((major*10 + minor) >= 2 && major < 1000 && minor < 1000){
-        fread(net->seen, sizeof(size_t), 1, fp);
+        //fread(net->seen, sizeof(size_t), 1, fp);
+        fread(net->seen, sizeof(uint64_t), 1, fp);
     } else {
         int iseen = 0;
         fread(&iseen, sizeof(int), 1, fp);
